+++ client/src/components/ri_components/OutfitList.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import OutfitItem from './OutfitItem.jsx';
-import { fitLeftScroll, fitRightScroll, addToOutfit } from './relatedHelperFunctions.js';
+import { fitLeftScroll, fitRightScroll, addToOutfit } from '../../helperFunctions/relatedHelperFunctions.js';
 
 const outfitList = ({ currentItem, setCurrentItemID, getStars }) => {
   const [outfitItemsIDs, setOutfitItemIDs] = useState([]);
