+++ client/src/helperFunctions/relatedHelperFunctions.js
